+++ Plugins/CrashRecovery/CrashRecovery.plugin.js
@@ -1,7 +1,6 @@
 //META{"name":"CrashRecovery","source":"https://github.com/1Lighty/BetterDiscordPlugins/blob/master/Plugins/CrashRecovery/","website":"https://1lighty.github.io/BetterDiscordStuff/?plugin=CrashRecovery","authorId":"239513071272329217","invite":"NYvWdN5","donate":"https://paypal.me/lighty13"}*//
 /*@cc_on
 @if (@_jscript)
-
   // Offer to self-install for clueless users that try to run this directly.
   var shell = WScript.CreateObject('WScript.Shell');
   var fs = new ActiveXObject('Scripting.FileSystemObject');
@@ -20,7 +19,6 @@
     shell.Popup('I\'m installed!\nJust go to settings, plugins and enable me!', 0, 'Successfully installed', 0x40);
   }
   WScript.Quit();
-
 @else@*/
 /*
  * Copyright © 2019-2020, _Lighty_
@@ -41,16 +39,20 @@ module.exports = (() => {
           twitter_username: ''
         }
       ],
-      version: '0.1.8',
+      version: '1.0.0',
       description: 'In the event that your Discord crashes, the plugin enables you to get Discord back to a working state, without needing to reload at all.',
       github: 'https://github.com/1Lighty',
       github_raw: 'https://raw.githubusercontent.com/1Lighty/BetterDiscordPlugins/master/Plugins/CrashRecovery/CrashRecovery.plugin.js'
     },
     changelog: [
       {
-        title: 'fixed',
-        type: 'fixed',
-        items: ['Crash fix.']
+        title: 'Added',
+        type: 'added',
+        items: ['Added extra info in console, mainly a cleaned up stack trace, decoded react errors and a component stack, to help plugin developers.', 'Powercord support has been added, plugin can now detect and potentially disable any misbehaving plugins that are causing the client to repeatedly crash.']
+      },
+      {
+        type: 'description',
+        content: 'Ty <@324622488644616195> (Juby210) for help on the Powercord side of things :heart:'
       }
     ],
     defaultConfig: [
@@ -65,7 +67,7 @@ module.exports = (() => {
   };
 
   /* Build */
-  const buildPlugin = ([Plugin, Api]) => {
+  const buildPlugin = ([Plugin, Api], BasePlugin) => {
     const { Logger, Utilities, WebpackModules, DiscordModules, Patcher, PluginUtilities, ReactTools, PluginUpdater } = Api;
     const { React, Dispatcher, FlexChild: Flex, GuildStore } = DiscordModules;
 
@@ -74,7 +76,17 @@ module.exports = (() => {
 
     const ModalStack = WebpackModules.getByProps('openModal');
 
-    return class CrashRecovery extends Plugin {
+    const isPowercord = !!window.powercord;
+    const BLACKLISTED_BUILTIN_PC_PLUGINS = ['Updater', 'Commands Manager', 'I18n', 'Module Manager', 'Settings']
+    const RE_PC_PLUGIN_NAME_FROM_PATH = /[\\\/]plugins[\\\/]([^\\\/]+)/;
+
+    const RE_INVARIANT = /error-decoder.html\?invariant=(\d+)([^\s]*)/;
+    const INVARIANTS_URL = 'https://raw.githubusercontent.com/facebook/react/master/scripts/error-codes/codes.json';
+    const ROOT_FOLDER = isPowercord ? window.powercord.basePath : BdApi.Plugins.folder;
+
+    const path = require('path');
+
+    return class CrashRecovery extends BasePlugin(Plugin) {
       constructor() {
         super();
         XenoLib.changeName(__filename, 'CrashRecovery');
@@ -110,6 +122,7 @@ module.exports = (() => {
           this.notificationId = null;
           if (this.disabledPlugins) XenoLib.Notifications.danger(`${this.disabledPlugins.map(e => e)} ${this.disabledPlugins.length > 1 ? 'have' : 'has'} been disabled to recover from the crash`, { timeout: 0 });
           if (this.suspectedPlugin) XenoLib.Notifications.danger(`${this.suspectedPlugin} ${this.suspectedPlugin2 !== this.suspectedPlugin && this.suspectedPlugin2 ? 'or ' + this.suspectedPlugin2 : ''} is suspected of causing the crash.`, { timeout: 10000 });
+          XenoLib.Notifications.info('Successfully recovered, more info can be found in the console (CTRL + SHIFT + I > console on top). Pass this information to support for further help.', { timeout: 10000 });
           this.disabledPlugins = null;
           this.suspectedPlugin = null;
           this.suspectedPlugin2 = null;
@@ -120,46 +133,132 @@ module.exports = (() => {
           if (BIVOverlay) appMount.append(BIVOverlay);
           Logger.info('Corrected incorrectly placed containers');
         });
-        this.patchAll();
+        Error.prepareStackTrace = (error, frames) => {
+          this._lastStackFrames = frames;
+          return error.stack
+        }
+        Utilities.suppressErrors(this.patchErrorBoundary)(this.promises.state);
+        if (!this.settings.lastErrorMapUpdate || Date.now() - this.settings.lastErrorMapUpdate > 2.628e+9) {
+          const https = require('https');
+          const req = https.request(INVARIANTS_URL, { headers: { 'origin': 'discord.com' } }, res => {
+            let body = '';
+            res.on('data', chunk => { body += chunk; });
+            res.on('end', () => {
+              if (res.statusCode !== 200) return;
+              try {
+                this.settings.errorMap = JSON.parse(body);
+                this.settings.lastErrorMapUpdate = Date.now();
+                this.saveSettings();
+              } catch { }
+            });
+          });
+          req.end();
+        }
       }
       onStop() {
         this.promises.state.cancelled = true;
         Patcher.unpatchAll();
         if (this.notificationId) XenoLib.Notifications.remove(this.notificationId);
+        delete Error.prepareStackTrace;
       }
 
-      queryResponsiblePlugins(stack) {
-        try {
-          const match = stack.match(/^.*(?:\\|\/|VM\d+ )([^\/\\\n]+)\.plugin\.js/gm);
-          const plugins = [];
-          if (!match || !match.length) return null;
-          let onDispatchEventFound = false;
-          for (let i = 0; i < match.length; i++) {
-            const pluginName = match[i].match(/(?:\\|\/)([^\/\\]+)\.plugin\.js/)[1];
-            if (match[i].indexOf('MessageLoggerV2.onDispatchEvent') !== -1) {
-              onDispatchEventFound = true;
-              continue;
-            }
-            if (onDispatchEventFound && /Object\.callback.*(?:\\|\/|VM\d+ )MessageLoggerV2\.plugin\.js/.test(match[i])) continue;
-            if (pluginName === '0PluginLibrary' || pluginName === this.name) continue;
 
-            let bbdplugin = null;
-            if (!BdApi.Plugins.get(pluginName)) {
-              /*
-               * go away Zack
-               */
-              for (const path in require.cache) {
-                const module = require.cache[path];
-                if (!module || !module.exports || !module.exports.plugin || module.exports.filename.indexOf(pluginName) !== 0) continue;
-                bbdplugin = module.exports;
-              }
-            }
-            const name = (bbdplugin && bbdplugin.name) || pluginName;
-            bbdplugin = null;
-            if (this.disabledPlugins && this.disabledPlugins.indexOf(name) !== -1) return { name: name };
-            plugins.push(name);
+      // Copyright (c) Facebook, Inc. and its affiliates
+      // https://github.com/facebook/react/blob/master/scripts/jest/setupTests.js#L171
+      decodeErrorMessage(message) {
+        if (!message) return message;
+
+        const [, invariant, argS] = message.match(RE_INVARIANT);
+        const code = parseInt(invariant, 10);
+        const args = argS
+          .split('&')
+          .filter(s => s.indexOf('args[]=') !== -1)
+          .map(s => s.substr('args[]='.length))
+          .map(decodeURIComponent);
+        const format = this.settings.errorMap[code];
+        if (!format) return message; // ouch
+        let argIndex = 0;
+        return format.replace(/%s/g, () => args[argIndex++]);
+      }
+
+      decodeStacks(stack, componentStack, baseComponentName = 'ErrorBoundary') {
+        if (!this.settings.errorMap) return { stack, componentStack };
+        if (RE_INVARIANT.test(stack)) stack = this.decodeErrorMessage(stack);
+        // Strip out Discord (and React) only functions
+        else stack = stack.split('\n')
+          .filter(l => l.indexOf('discordapp.com/assets') === -1 && l.indexOf('discord.com/assets') === -1)
+          .join('\n')
+          .split(ROOT_FOLDER) // transform paths to relative
+          .join('');
+
+        // Only show up to the error boundary
+        const splitComponentStack = componentStack.split('\n').filter(e => e);
+        const stackEnd = splitComponentStack.findIndex(l => l.indexOf(`in ${baseComponentName}`) !== -1);
+        if (stackEnd !== -1 && baseComponentName) splitComponentStack.splice(stackEnd + 1, splitComponentStack.length);
+        componentStack = splitComponentStack.join('\n');
+        return { stack, componentStack };
+      }
+
+      /*
+        MUST return either an array with the plugin name
+        or a string of the plugin name
+       */
+      queryResponsiblePlugins() {
+        try {
+          const stack = this._bLastStackFrames
+            // filter out blank functions (like from console or whatever)
+            .filter(e => e.getFileName() || e.getFunctionName() || e.getMethodName())
+            // filter out discord functions
+            .filter(e => e.getFileName().indexOf(ROOT_FOLDER) !== -1)
+            // convert CallSites to only useful info
+            .map(e => ({ filename: e.getFileName(), functionName: e.getFunctionName() || e.getMethodName() }))
+            // filter out ZeresPluginLibrary and ourselves
+            .filter(({ filename }) => filename.lastIndexOf('0PluginLibrary.plugin.js') !== filename.length - 24 && filename.lastIndexOf(`${this.name}.plugin.js`) !== filename.length - (this.name.length + 10));
+          // Filter out MessageLoggerV2 dispatch patch, is a 2 part step
+          for (let i = 0, len = stack.length; i < len; i++) {
+            const { filename, functionName } = stack[i];
+            if (filename.lastIndexOf('MessageLoggerV2.plugin.js') !== filename.length - 25) continue;
+            if (functionName !== 'onDispatchEvent') continue;
+            if (stack[i + 1].functionName !== 'callback') break;
+            stack.splice(i, 2);
+            break;
           }
-          return plugins;
+          const plugins = stack.map(({ filename }) => {
+            try {
+              const bdname = path.basename(filename);
+              if (bdname.indexOf('.plugin.js') === bdname.length - 10) {
+                const [pluginName] = bdname.split('.plugin.js');
+                if (BdApi.Plugins.get(pluginName)) return pluginName;
+                /*
+                 * go away Zack
+                 */
+                for (const path in require.cache) {
+                  const module = require.cache[path];
+                  if (!module || !module.exports || !module.exports.plugin || module.exports.filename.indexOf(bdname) !== 0) continue;
+                  return module.exports.id;
+                }
+                return null;
+              }
+              else if (isPowercord) {
+                const [, pcname] = RE_PC_PLUGIN_NAME_FROM_PATH.exec(filename);
+                const { pluginManager } = window.powercord;
+                const plugin = pluginManager.get(pcname);
+                if (!plugin) return null;
+                const { name } = plugin.manifest;
+                if (BLACKLISTED_BUILTIN_PC_PLUGINS.indexOf(name) !== -1) return null;
+                return pcname;
+              }
+            } catch (err) {
+              Logger.stacktrace('Error fetching plugin')
+            }
+          }).filter((name, idx, self) => name && self.indexOf(name) === idx);
+          const ret = [];
+          for (let i = 0, len = plugins.length; i < len; i++) {
+            const name = plugins[i];
+            if (this.disabledPlugins && this.disabledPlugins.indexOf(name) !== -1) return name;
+            ret.push(name);
+          }
+          return ret;
         } catch (e) {
           Logger.stacktrace('query error', e);
           return null;
@@ -202,7 +301,14 @@ module.exports = (() => {
         });
       }
 
-      handleCrash(_this, stack, isRender) {
+      handleCrash(_this, stack, componentStack, isRender = false) {
+        this._bLastStackFrames = this._lastStackFrames;
+        try {
+          const decoded = this.decodeStacks(stack, componentStack);
+          Logger.error('HEY OVER HERE! Show this to the plugin developer or in the support server!\nPrettified stacktraces, stack:\n', decoded.stack, '\nComponent stack:\n', decoded.componentStack);
+        } catch (err) {
+          Logger.stacktrace('Failed decoding stack!', err);
+        }
         this.onCrashRecoveredDelayedCall.cancel();
         if (!isRender) {
           _this.setState({
@@ -212,9 +318,9 @@ module.exports = (() => {
         if (!this.notificationId) {
           this.notificationId = XenoLib.Notifications.danger('Crash detected, attempting recovery', { timeout: 0, loading: true });
         }
-        const responsiblePlugins = this.queryResponsiblePlugins(stack);
+        const responsiblePlugins = this.queryResponsiblePlugins();
         if (responsiblePlugins && !Array.isArray(responsiblePlugins)) {
-          XenoLib.Notifications.update(this.notificationId, { content: `Failed to recover from crash, ${responsiblePlugins.name} is not stopping properly`, loading: false });
+          XenoLib.Notifications.update(this.notificationId, { content: `Failed to recover from crash, ${responsiblePlugins} is not stopping properly`, loading: false });
           return;
         }
         if (!this.attempts) {
@@ -231,12 +337,14 @@ module.exports = (() => {
           Dispatcher.wait(() => DiscordModules.NavigationUtils.transitionTo('/channels/@me'));
           XenoLib.Notifications.update(this.notificationId, { content: `Failed, switching channels` });
         } else if (this.attempts >= 2) {
+          const [name] = responsiblePlugins;
           try {
-            BdApi.Plugins.disable(responsiblePlugins[0]);
+            if (BdApi.Plugins.get(name)) BdApi.Plugins.disable(name);
+            else if (isPowercord) window.powercord.pluginManager.disable(name);
           } catch (e) { }
-          XenoLib.Notifications.update(this.notificationId, { content: `Failed, suspecting ${responsiblePlugins[0]} for recovery failure` });
+          XenoLib.Notifications.update(this.notificationId, { content: `Failed, suspecting ${name} for recovery failure` });
           if (!this.disabledPlugins) this.disabledPlugins = [];
-          this.disabledPlugins.push(responsiblePlugins[0]);
+          this.disabledPlugins.push(name);
         }
         this.setStateTimeout = setTimeout(() => {
           this.setStateTimeout = null;
@@ -251,19 +359,15 @@ module.exports = (() => {
 
       /* PATCHES */
 
-      patchAll() {
-        this.patchErrorBoundary(this.promises.state);
-      }
-
       patchErrorBoundary() {
         const ErrorBoundary = WebpackModules.getByDisplayName('ErrorBoundary');
-        Patcher.instead(ErrorBoundary.prototype, 'componentDidCatch', (_this, [{ message, stack }, { componentStack }], orig) => {
-          this.handleCrash(_this, stack);
+        Patcher.instead(ErrorBoundary.prototype, 'componentDidCatch', (_this, [{ stack }, { componentStack }], orig) => {
+          this.handleCrash(_this, stack, componentStack);
         });
         Patcher.after(ErrorBoundary.prototype, 'render', (_this, _, ret) => {
           if (!_this.state.error) return;
           if (!this.notificationId) {
-            this.handleCrash(_this, _this.state.error.stack, true);
+            this.handleCrash(_this, _this.state.error.stack, _this.state.info.componentStack, true);
           }
           /* better be safe than sorry! */
           if (!_this.state.customPageError) {
@@ -327,193 +431,103 @@ module.exports = (() => {
         ErrorBoundaryInstance.state.customPageError = false;
         ErrorBoundaryInstance.forceUpdate();
       }
-
-      /* PATCHES */
-
-      showChangelog(footer) {
-        XenoLib.showChangelog(`${this.name} has been updated!`, this.version, this._config.changelog);
-      }
-
-      getSettingsPanel() {
-        return this.buildSettingsPanel().getElement();
-      }
-
-      get [Symbol.toStringTag]() {
-        return 'Plugin';
-      }
-      get css() {
-        return this._css;
-      }
-      get name() {
-        return config.info.name;
-      }
-      get short() {
-        let string = '';
-
-        for (let i = 0, len = config.info.name.length; i < len; i++) {
-          const char = config.info.name[i];
-          if (char === char.toUpperCase()) string += char;
-        }
-
-        return string;
-      }
-      get author() {
-        return config.info.authors.map(author => author.name).join(', ');
-      }
-      get version() {
-        return config.info.version;
-      }
-      get description() {
-        return config.info.description;
-      }
+      showChangelog = () => XenoLib.showChangelog(`${this.name} has been updated!`, this.version, this._config.changelog);
+      getSettingsPanel = () =>
+        this.buildSettingsPanel()
+          .append(new XenoLib.Settings.PluginFooter(() => this.showChangelog()))
+          .getElement();
     };
   };
 
   /* Finalize */
 
+  /* shared getters */
+  const BasePlugin = cl =>
+    class extends cl {
+      constructor() {
+        super();
+        Object.defineProperties(this, {
+          name: { get: () => config.info.name },
+          short: { get: () => config.info.name.split('').reduce((acc, char) => acc + (char === char.toUpperCase() ? char : '')) },
+          author: { get: () => config.info.authors.map(author => author.name).join(', ') },
+          version: { get: () => config.info.version },
+          description: { get: () => config.info.description }
+        });
+      }
+    };
+
+  /* this new lib loader is lit */
   let ZeresPluginLibraryOutdated = false;
   let XenoLibOutdated = false;
   try {
     if (global.BdApi && 'function' == typeof BdApi.getPlugin) {
-      const i = (i, n) => ((i = i.split('.').map(i => parseInt(i))), (n = n.split('.').map(i => parseInt(i))), !!(n[0] > i[0]) || !!(n[0] == i[0] && n[1] > i[1]) || !!(n[0] == i[0] && n[1] == i[1] && n[2] > i[2])),
-        n = (n, e) => n && n._config && n._config.info && n._config.info.version && i(n._config.info.version, e),
-        e = BdApi.getPlugin('ZeresPluginLibrary'),
-        o = BdApi.getPlugin('XenoLib');
-      n(e, '1.2.23') && (ZeresPluginLibraryOutdated = !0), n(o, '1.3.26') && (XenoLibOutdated = !0);
+      const a = (c, a) => ((c = c.split('.').map(b => parseInt(b))), (a = a.split('.').map(b => parseInt(b))), !!(a[0] > c[0])) || !!(a[0] == c[0] && a[1] > c[1]) || !!(a[0] == c[0] && a[1] == c[1] && a[2] > c[2]),
+        b = (b, c) => ((b && b._config && b._config.info && b._config.info.version && a(b._config.info.version, c)) || typeof global.isTab !== 'undefined'),
+        c = BdApi.getPlugin('ZeresPluginLibrary'),
+        d = BdApi.getPlugin('XenoLib');
+      b(c, '1.2.26') && (ZeresPluginLibraryOutdated = !0), b(d, '1.3.31') && (XenoLibOutdated = !0);
     }
-  } catch (i) {
-    console.error('Error checking if libraries are out of date', i);
+  } catch (a) {
+    console.error('Error checking if libraries are out of date', a);
   }
 
-  return !global.ZeresPluginLibrary || !global.XenoLib || ZeresPluginLibraryOutdated || XenoLibOutdated
-    ? class {
+  /* to anyone asking "why are you checking if x is out of date", well you see, sometimes, for whatever reason
+     the libraries are sometimes not updating for people. Either it doesn't check for an update, or the request
+     for some odd reason just fails. Yet, plugins update just fine with the same domain.
+   */
+  return !global.ZeresPluginLibrary || !global.XenoLib || global.DiscordJS || ZeresPluginLibraryOutdated || XenoLibOutdated
+    ? class extends BasePlugin(class { }) {
       constructor() {
+        super();
         this._XL_PLUGIN = true;
+        this.getName = () => this.name.replace(/\s+/g, '');
+        this.getAuthor = () => this.author;
+        this.getVersion = () => this.version;
+        this.getDescription = () => this.description + global.BetterDiscordConfig ? '' : ' You are missing libraries for this plugin, please enable the plugin and click Download Now.';
         this.start = this.load = this.handleMissingLib;
-      }
-      getName() {
-        return this.name.replace(/\s+/g, '');
-      }
-      getAuthor() {
-        return this.author;
-      }
-      getVersion() {
-        return this.version;
-      }
-      getDescription() {
-        return this.description + ' You are missing libraries for this plugin, please enable the plugin and click Download Now.';
       }
       start() { }
       stop() { }
       handleMissingLib() {
-        const a = BdApi.findModuleByProps('openModal', 'hasModalOpen');
-        if (a && a.hasModalOpen(`${this.name}_DEP_MODAL`)) return;
-        const b = !global.XenoLib,
-          c = !global.ZeresPluginLibrary,
-          d = (b && c) || ((b || c) && (XenoLibOutdated || ZeresPluginLibraryOutdated)),
-          e = (() => {
-            let a = '';
-            return b || c ? (a += `Missing${XenoLibOutdated || ZeresPluginLibraryOutdated ? ' and outdated' : ''} `) : (XenoLibOutdated || ZeresPluginLibraryOutdated) && (a += `Outdated `), (a += `${d ? 'Libraries' : 'Library'} `), a;
-          })(),
-          f = (() => {
-            let a = `The ${d ? 'libraries' : 'library'} `;
-            return b || XenoLibOutdated ? ((a += 'XenoLib '), (c || ZeresPluginLibraryOutdated) && (a += 'and ZeresPluginLibrary ')) : (c || ZeresPluginLibraryOutdated) && (a += 'ZeresPluginLibrary '), (a += `required for ${this.name} ${d ? 'are' : 'is'} ${b || c ? 'missing' : ''}${XenoLibOutdated || ZeresPluginLibraryOutdated ? (b || c ? ' and/or outdated' : 'outdated') : ''}.`), a;
-          })(),
-          g = BdApi.findModuleByDisplayName('Text'),
-          h = BdApi.findModuleByDisplayName('ConfirmModal'),
-          i = () => BdApi.alert(e, BdApi.React.createElement('span', {}, BdApi.React.createElement('div', {}, f), `Due to a slight mishap however, you'll have to download the libraries yourself. This is not intentional, something went wrong, errors are in console.`, c || ZeresPluginLibraryOutdated ? BdApi.React.createElement('div', {}, BdApi.React.createElement('a', { href: 'https://betterdiscord.net/ghdl?id=2252', target: '_blank' }, 'Click here to download ZeresPluginLibrary')) : null, b || XenoLibOutdated ? BdApi.React.createElement('div', {}, BdApi.React.createElement('a', { href: 'https://betterdiscord.net/ghdl?id=3169', target: '_blank' }, 'Click here to download XenoLib')) : null));
-        if (!a || !h || !g) return console.error(`Missing components:${(a ? '' : ' ModalStack') + (h ? '' : ' ConfirmationModalComponent') + (g ? '' : 'TextElement')}`), i();
-        class j extends BdApi.React.PureComponent {
+        if ("undefined" != typeof global.isTab) return;
+        const a = !!window.powercord && -1 !== (window.bdConfig && window.bdConfig.dataPath || "").indexOf("bdCompat") && "function" == typeof BdApi.__getPluginConfigPath, b = BdApi.findModuleByProps("openModal", "hasModalOpen");
+        if (b && b.hasModalOpen(`${this.name}_DEP_MODAL`)) return;
+        const c = !global.XenoLib, d = !global.ZeresPluginLibrary, e = c && d || (c || d) && (XenoLibOutdated || ZeresPluginLibraryOutdated),
+          f = (() => { let a = ""; return c || d ? a += `Missing${XenoLibOutdated || ZeresPluginLibraryOutdated ? " and outdated" : ""} ` : (XenoLibOutdated || ZeresPluginLibraryOutdated) && (a += `Outdated `), a += `${e ? "Libraries" : "Library"} `, a })(),
+          g = (() => {
+            let a = `The ${e ? "libraries" : "library"} `; return c || XenoLibOutdated ? (a += "XenoLib ", (d || ZeresPluginLibraryOutdated) && (a += "and ZeresPluginLibrary ")) : (d || ZeresPluginLibraryOutdated) && (a += "ZeresPluginLibrary "),
+              a += `required for ${this.name} ${e ? "are" : "is"} ${c || d ? "missing" : ""}${XenoLibOutdated || ZeresPluginLibraryOutdated ? c || d ? " and/or outdated" : "outdated" : ""}.`, a
+          })(), h = BdApi.findModuleByDisplayName("Text"), i = BdApi.findModuleByDisplayName("ConfirmModal"),
+          j = () => BdApi.alert(f, BdApi.React.createElement("span", { style: { color: "white" } }, BdApi.React.createElement("div", {}, g), `Due to a slight mishap however, you'll have to download the libraries yourself. This is not intentional, something went wrong, errors are in console.`,
+            d || ZeresPluginLibraryOutdated ? BdApi.React.createElement("div", {}, BdApi.React.createElement("a", { href: "https://betterdiscord.net/ghdl?id=2252", target: "_blank" }, "Click here to download ZeresPluginLibrary")) : null, c || XenoLibOutdated ? BdApi.React.createElement("div", {},
+              BdApi.React.createElement("a", { href: "https://betterdiscord.net/ghdl?id=3169", target: "_blank" }, "Click here to download XenoLib")) : null)); if (global.ohgodohfuck) return; if (!b || !i || !h) return console.error(`Missing components:${(b ? "" : " ModalStack") + (i ? "" : " ConfirmationModalComponent") + (h ? "" : "TextElement")}`),
+                j(); class k extends BdApi.React.PureComponent {
           constructor(a) {
-            super(a), (this.state = { hasError: !1 }), (this.componentDidCatch = a => (console.error(`Error in ${this.props.label}, screenshot or copy paste the error above to Lighty for help.`), this.setState({ hasError: !0 }), 'function' == typeof this.props.onError && this.props.onError(a))), (this.render = () => (this.state.hasError ? null : this.props.children));
+            super(a), this.state = { hasError: !1 }, this.componentDidCatch = a => (console.error(`Error in ${this.props.label}, screenshot or copy paste the error above to Lighty for help.`), this.setState({ hasError: !0 }), "function" == typeof this.props.onError && this.props.onError(a)),
+              this.render = () => this.state.hasError ? null : this.props.children
           }
-        }
-        let k = !1,
-          l = !1;
-        const m = a.openModal(
-          b => {
-            if (l) return null;
-            try {
-              return BdApi.React.createElement(
-                j,
-                { label: 'missing dependency modal', onError: () => (a.closeModal(m), i()) },
-                BdApi.React.createElement(
-                  h,
-                  Object.assign(
-                    {
-                      header: e,
-                      children: BdApi.React.createElement(g, { size: g.Sizes.SIZE_16, children: [`${f} Please click Download Now to download ${d ? 'them' : 'it'}.`] }),
-                      red: !1,
-                      confirmText: 'Download Now',
-                      cancelText: 'Cancel',
-                      onCancel: b.onClose,
-                      onConfirm: () => {
-                        if (k) return;
-                        k = !0;
-                        const b = require('request'),
-                          c = require('fs'),
-                          d = require('path'),
-                          e = BdApi.Plugins && BdApi.Plugins.folder ? BdApi.Plugins.folder : window.ContentManager.pluginsFolder,
-                          f = () => {
-                            (global.XenoLib && !XenoLibOutdated) ||
-                              b('https://raw.githubusercontent.com/1Lighty/BetterDiscordPlugins/master/Plugins/1XenoLib.plugin.js', (b, f, g) => {
-                                try {
-                                  if (b || 200 !== f.statusCode) return a.closeModal(m), i();
-                                  c.writeFile(d.join(e, '1XenoLib.plugin.js'), g, () => { });
-                                } catch (b) {
-                                  console.error('Fatal error downloading XenoLib', b), a.closeModal(m), i();
-                                }
-                              });
-                          };
-                        !global.ZeresPluginLibrary || ZeresPluginLibraryOutdated
-                          ? b('https://raw.githubusercontent.com/rauenzi/BDPluginLibrary/master/release/0PluginLibrary.plugin.js', (b, g, h) => {
-                            try {
-                              if (b || 200 !== g.statusCode) return a.closeModal(m), i();
-                              c.writeFile(d.join(e, '0PluginLibrary.plugin.js'), h, () => { }), f();
-                            } catch (b) {
-                              console.error('Fatal error downloading ZeresPluginLibrary', b), a.closeModal(m), i();
-                            }
-                          })
-                          : f();
-                      }
-                    },
-                    b,
-                    { onClose: () => { } }
-                  )
-                )
-              );
-            } catch (b) {
-              return console.error('There has been an error constructing the modal', b), (l = !0), a.closeModal(m), i(), null;
-            }
-          },
-          { modalKey: `${this.name}_DEP_MODAL` }
-        );
-      }
-      get [Symbol.toStringTag]() {
-        return 'Plugin';
-      }
-      get name() {
-        return config.info.name;
-      }
-      get short() {
-        let string = '';
-        for (let i = 0, len = config.info.name.length; i < len; i++) {
-          const char = config.info.name[i];
-          if (char === char.toUpperCase()) string += char;
-        }
-        return string;
-      }
-      get author() {
-        return config.info.authors.map(author => author.name).join(', ');
-      }
-      get version() {
-        return config.info.version;
-      }
-      get description() {
-        return config.info.description;
+        } let l = !!global.DiscordJS, m = !1; const n = b.openModal(c => {
+          if (m) return null; try {
+            return BdApi.React.createElement(k, { label: "missing dependency modal", onError: () => (b.closeModal(n), j()) }, BdApi.React.createElement(i,
+              Object.assign({
+                header: f, children: BdApi.React.createElement(h, { size: h.Sizes.SIZE_16, children: [`${g} Please click Download Now to download ${e ? "them" : "it"}.`] }), red: !1, confirmText: "Download Now", cancelText: "Cancel", onCancel: c.onClose, onConfirm: () => {
+                  if (l) return; l = !0; const c = require("request"), d = require("fs"),
+                    e = require("path"), f = BdApi.Plugins && BdApi.Plugins.folder ? BdApi.Plugins.folder : window.ContentManager.pluginsFolder, g = () => {
+                      global.XenoLib && !XenoLibOutdated || c("https://raw.githubusercontent.com/1Lighty/BetterDiscordPlugins/master/Plugins/1XenoLib.plugin.js", (c, g, h) => {
+                        try {
+                          if (c || 200 !== g.statusCode) return b.closeModal(n),
+                            j(); d.writeFile(e.join(f, "1XenoLib.plugin.js"), h, () => { BdApi.isSettingEnabled("fork-ps-5") && !a || BdApi.Plugins.reload(this.getName()) })
+                        } catch (a) { console.error("Fatal error downloading XenoLib", a), b.closeModal(n), j() }
+                      })
+                    }; !global.ZeresPluginLibrary || ZeresPluginLibraryOutdated ? c("https://raw.githubusercontent.com/rauenzi/BDPluginLibrary/master/release/0PluginLibrary.plugin.js",
+                      (a, c, h) => { try { if (a || 200 !== c.statusCode) return b.closeModal(n), j(); d.writeFile(e.join(f, "0PluginLibrary.plugin.js"), h, () => { }), g() } catch (a) { console.error("Fatal error downloading ZeresPluginLibrary", a), b.closeModal(n), j() } }) : g()
+                }
+              }, c, { onClose: () => { } })))
+          } catch (a) { return console.error("There has been an error constructing the modal", a), m = !0, b.closeModal(n), j(), null }
+        }, { modalKey: `${this.name}_DEP_MODAL` });
       }
     }
-    : buildPlugin(global.ZeresPluginLibrary.buildPlugin(config));
+    : buildPlugin(global.ZeresPluginLibrary.buildPlugin(config), BasePlugin);
 })();
 
 /*@end@*/